requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements.in
#
biopython==1.79
    # via -r requirements.in
cycler==0.10.0
    # via matplotlib
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.3.4
    # via -r requirements.in
numpy==1.19.5
    # via
    #   -r requirements.in
    #   biopython
    #   matplotlib
pillow==8.3.2
    # via matplotlib
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.2
    # via matplotlib
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
